Refactor google drive interface to not need gid pgid and rename packages to be more reflective of purpose
